swap out old addressing files
