--- assembly/cpu/addr/index.ts
@@ -1,7 +1,0 @@
-
-export { addr_absolute } from './absolute';
-export { addr_blockMove } from './block-move';
-export { addr_directPage } from './direct-page';
-export { addr_immediate } from './immediate';
-export { addr_programCounterRelative } from './program-counter-relative';
-export { addr_stackRelative } from './stack-relative';
