+++ assembly/cpu/addressing/index.ts
@@ -1,27 +1,7 @@
 
-export { addr_immediate_u8, addr_immediate_u16, addr_immediate_u24 } from './immediate';
-
-export {
-	addr_absolute,
-	addr_absoluteIndexedX, addr_absoluteIndexedY,
-	addr_absoluteLong,
-	addr_absoluteLongIndexedX, addr_absoluteLongIndexedY
-} from './absolute';
-
-export {
-	addr_directPage,
-	addr_directPageIndexedX, addr_directPageIndexedY,
-	addr_directPageIndirect,
-	addr_directPageIndexedIndirectX, addr_directPageIndirectIndexedY,
-	addr_directPageIndirectLong,
-	addr_directPageIndirectLongIndexedX, addr_directPageIndirectLongIndexedY
-} from './direct-page';
-
-// export { addr_blockMove } from './block-move';
-
-export { addr_programCounterRelative, addr_programCounterRelativeLong } from './program-counter-relative';
-
-export { addr_stackRelative, addr_stackRelativeIndirectIndexedX, addr_stackRelativeIndirectIndexedY } from './stack-relative';
-
-// TODO: Do address modes actually impart deterministic cycle cost increases? Can
-// the be moved into these files?
+export { addr_absolute } from './absolute';
+export { addr_blockMove } from './block-move';
+export { addr_directPage } from './direct-page';
+export { addr_immediate } from './immediate';
+export { addr_programCounterRelative } from './program-counter-relative';
+export { addr_stackRelative } from './stack-relative';
